// src/grids/src/grid/base/interface.ts
export interface NotifyArgs {
    /** Defines the total records. */
    records?: Object[];
    /** Defines the record count. */
    count?: number;
    /** Defines the request type. */
    requestType?: Action;
    /** Defines the module. */
    module?: string;
    /** Defines the enable property. */
    enable?: boolean;
    /** Defines the properties. */
    properties?: Object;
    /** Defines the virtualization info like block, event name and next page need to be loaded. */
    virtualInfo?: VirtualInfo;
    /** Defines whether the action needs to be cancel or not. */
    cancel?: boolean;
    /** Defines the rows. */
    rows?: Row<Column>[];
    /** Defines whether the grid is frozen or not. */
    isFrozen?: boolean;
    /** Defines the arguments. */
    args?: NotifyArgs;
    /** Defines the scroll top value. */
    scrollTop?: Object;
    /** Defines the old properties. */
    oldProperties?: string[];
    /** Defines the focus element. */
    focusElement?: HTMLElement;
    /** Defines the row object. */
    rowObject?: Row<Column>;
    /** Defines the movable content to be rendered. */
    renderMovableContent?: boolean;
    /** Defines the frozen right content. */
    renderFrozenRightContent?: boolean;
    /** Defines the promise. */
    promise?: Promise<Object>;
    /** Defines the frozen rows are rendered or not. */
    isFrozenRowsRender?: boolean;
    /** Defines the action. */
    action?: string;
    /** Defines the searched value. */
    searchString?: string;
    /** Defines the focus on Group. */
    preventFocusOnGroup?: boolean;
}
// ...
export interface IModelGenerator<T> {
    generateRows(data: Object, args?: Object): Row<T>[];
    refreshRows?(input?: Row<T>[]): Row<T>[];
}
// ...
export interface VirtualInfo {
    data?: boolean;
    event?: string;
    block?: number;
    page?: number;
    currentPage?: number;
    direction?: string;
    blockIndexes?: number[];
    columnIndexes?: number[];
    columnBlocks?: number[];
    loadSelf?: boolean;
    loadNext?: boolean;
    nextInfo?: { page?: number };
    sentinelInfo?: SentinelType;
    offsets?: Offsets;
    startIndex?: number;
    endIndex?: number;
}